codenano-service/pyproject.toml
[project]
name = "cc-agent-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sse-starlette>=3.3.4",
    "uvicorn[standard]>=0.44.0",
]
[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
